==> 1-extract-teacher-assessments.ipynb <==
import pandas as pd

# Read original annotation data
data = pd.read_excel('/vol/tensusers5/wharmsen/astla-data/dart-preposttest/data-in-different-formats/1-input-files/DART_test_data.xlsx')
data.loc[0,:]

# Get names of audio files that all 51 teachers annotated
audioFileFrequencies = data['Audio recording'].value_counts()
mostAnnotatedAudioFiles = audioFileFrequencies[0:6].index
mostAnnotatedAudioFiles

# Check if there are 51 unique teacherIDs of each audio file.
for audioFile in mostAnnotatedAudioFiles:
    selection = data[data['Audio recording'] == audioFile]
    uniqueTeacherIDs = list(set(list(selection['Teacher'])))
    print(audioFile.split('/')[-1], len(uniqueTeacherIDs))

#Print specifics of audio file
output = []
for audioFile in mostAnnotatedAudioFiles:
    selection = data[data['Audio recording'] == audioFile].reset_index()
    kern = selection.loc[0, 'Kern']
    exerciseID = selection.loc[0, 'ExerciseID']
    test = kern + str(exerciseID)
    studentID = selection.loc[0, 'StudentID']
    nrUniqueTeachers = len(list(set(list(selection['Teacher']))))
    nrUniqueWords = len(list(set(list(selection['Woord']))))
    output.append([audioFile.split('/')[-1], kern, exerciseID, test, studentID, nrUniqueTeachers, nrUniqueWords])

audioMetaData = pd.DataFrame(output, columns=['audio', 'kern', 'exerciseID', 'test', 'speaker', 'nrOfAnnotations', 'nrOfWords'])
audioMetaData

# Read prompts
references = pd.read_csv("../../astla-data/dart-preposttest/data-in-different-formats/1-input-files/reference_wordlist.csv", sep=";")
references = references.set_index('name')
references

allAssessmentDF = pd.DataFrame()
for audioFile in mostAnnotatedAudioFiles:
    selection = data[data['Audio recording'] == audioFile].reset_index()
    uniqueTeachers = list(set(list(selection['Teacher'])))

    # Get Test Name
    audio = audioFile.split('/')[-1]
    testName = audioMetaData.set_index('audio').loc[audio,'test']

    # Get prompt wordlist
    uniqueWords = references.loc[testName, 'prompt'].split(' ')
    index = [audio.replace('.mp3', '') + '_' + word for word in uniqueWords]

    # Create dataframe to save teacher judgements
    assessmentDF = pd.DataFrame(columns = list(set(list(selection['Teacher']))), 
                   index = index)
    
    for idx,row in selection.iterrows():
        teacherID = row['Teacher']
        index = audio.replace('.mp3', '') + '_' + row['Woord']
        assessment = row['Correct']

        assessmentDF.loc[index, teacherID] = assessment

    assessmentDF.to_csv('../../astla-data/dart-preposttest/slate-data/1-teacherAssessments/' + audio.replace('.mp3', '') + '_' + testName + '_teacherAssessments.csv', sep=',', encoding='utf-8')
    allAssessmentDF = pd.concat([allAssessmentDF, assessmentDF])

allAssessmentDF.to_csv('../../astla-data/dart-preposttest/slate-data/1-teacherAssessments/' + 'all_teacherAssessments.csv', sep=',', encoding='utf-8')
allAssessmentDF
