==> 3-add-whisper-t-dis.ipynb <==
import json
import glob
import pandas as pd
import reversedadapt.getAlignmentsPosttestDart as reversedadapt
import reversedadapt.textToParts as adapttexttoparts

# Read all json whisper results, and save the six relevant ones in whisperDict

jsonFiles = glob.glob('/vol/tensusers5/wharmsen/astla-data/dart-preposttest/slate-data/2-whisper-t-dis/*json')

# All SLaTE results in Dict
whisperDict = {}
for jsonFile in jsonFiles:
    with open(jsonFile, 'r') as f:
        data = json.load(f)
        fileName = jsonFile.split('/')[-1].replace('.json', '')
        whisperDict[fileName] = data

print(whisperDict.keys())

# Create Whisper Transcription Dict

whisperTranscriptionDict = {}
for file in whisperDict.keys():
    data = whisperDict[file]
    wordDict = {}
    for segment in data['segments']:
        words = segment['words']
        for word in words:

            label = word['text']

            # If no dislefluency
            if(label != '[*]'):

                start = word['start']
                end = word['end']
                confidence = word['confidence']

                # Save word information in WordDict
                wordDict[label] = {
                    'confidence': confidence,
                    'start': start,
                    'end': end
                    }
    
    whisperTranscriptionDict[file] = wordDict

print(whisperTranscriptionDict['09eae7df-db7c-491b-b7c4-a7c9fcf62f47']['jong'])

#df = pd.DataFrame(whisperOutputList, columns=['wh_word', 'wh_acc', 'wh_start', 'wh_end'])
#df.to_csv('../../astla-data/dart-preposttest/slate-data/5-asr-analysis-results/whisper-word-assessments.csv', index=False)
            


# For each word in the prompt, we need to extract a correct/incorrect assessment from the Whisper ASR-output.
# We will compute this assessment in two ways:
# 1) Align prompt and output
# 2) For each prompt word, check whether it is in the output

# Prompt word is incorrectly pronounced when it is not in the ASR output, or is in the ASR-output, but has a very low confidence score


references = pd.read_csv("../../astla-data/dart-preposttest/data-in-different-formats/1-input-files/reference_wordlist.csv", sep=";")
references = references.set_index('name')
references

filelist = glob.glob("/vol/tensusers5/wharmsen/astla-data/dart-preposttest/slate-data/1-teacherAssessments/*.csv")

fileTestDict = {}
for file in filelist:
    parsedName = file.split('/')[-1].split('_')
    if(len(parsedName)==3):
        audio = parsedName[0]
        test = parsedName[1]
        fileTestDict[audio] = test

# Source: alignments.py
import helper_scripts.aligner as aligner

def getPrompt(audio):
    testType = fileTestDict[audio]
    return references.loc[testType, 'prompt']

def checkIfPromptInAsrTransAligned(prompt, asrTransAligned):
    return prompt in asrTransAligned


testDict = [
### false start, substitutions, substitutions
{
    'prompt': "jong lach zieke",
    'asrTrans': "jon jong lang ring"
},

### false start, 3/4 attempt correct, false start
{
    'prompt': "jong onze zieke",
    'asrTrans': "jon jong onzin onz onze z zeker"
},
### second word not pronounced
{
    'prompt': "jong onze zieke",
    'asrTrans': "jong zieke"
},
### second word not pronounced
{
    'prompt': "jong onze zieke",
    'asrTrans': "jon jong z zieke zeker"
}]

for inputPair in testDict:

    reversedadapt.getAlignmentsPromptAsrTrans(inputPair['prompt'], inputPair['asrTrans'])


        # Add begin, end and confidence scores

# Initialize whisperOutputDF
whisperOutputDF = pd.DataFrame(columns=[
                               'aligned_asrTrans', 'reversed_aligned_asrTrans', 'correct', 'whisperWordIndex', 'confidence', 'begin', 'end', 'filename'])

# For each whisper output file, extract relevant information and add them to whisperOutputDF
for file in whisperTranscriptionDict.keys():
    
    print("Processing ", file, "...")

    # Get ASR transcription and prompt
    asrTranscription = " ".join(whisperTranscriptionDict[file])
    prompt = getPrompt(file)

    # Align the ASR transcription and prompt using reversed ADAPT & non-reversed ADAPT strategy with textToParts to speed up the process
    target_text_parts_list, original_text_parts_list = adapttexttoparts.textToParts(
        prompt.lower(), asrTranscription.lower())
    
    alignmentDFList= []
    for part_idx, part in enumerate(target_text_parts_list):
        alignmentPartDF = reversedadapt.getAlignmentsPromptAsrTrans(
        target_text_parts_list[part_idx].lower(), original_text_parts_list[part_idx].lower())
        alignmentDFList.append(alignmentPartDF)

    alignmentDF = pd.concat(alignmentDFList)
    # alignmentDF = reversedadapt.getAlignmentsPromptAsrTrans(
    #     prompt.lower(), asrTranscription.lower())
    alignmentDF.to_csv(
        '../../astla-data/dart-preposttest/slate-data/2-whisper-t-dis/'+file+'.csv')

    # Match info from alignmentDF with confidence scores, begin and end times

    # Try to find part of words in whisper output that match the prompt. 
    # priorities, first aligned_asrTrans, then prompt, then reversed_aligned_asrTrans
    foundIndexTermsList = []
    confidenceList = []
    startTimeList = []
    endTimeList = []
    for idx, row in alignmentDF.iterrows():
        prompt = row.name
        aligned_asrTrans = row['aligned_asrTrans'].replace('-', '')
        reversed_aligned_asrTrans = row['reversed_aligned_asrTrans'].replace(
            '-', '')
        
        indexWordAlignedAsrTrans = aligned_asrTrans if aligned_asrTrans in whisperTranscriptionDict[file].keys() else '' 
        indexWordPrompt = prompt if prompt in whisperTranscriptionDict[file].keys() else ''
        indexWordReversedAlignedAsrTrans = reversed_aligned_asrTrans if reversed_aligned_asrTrans in whisperTranscriptionDict[file].keys() else '' 
        
        if indexWordAlignedAsrTrans != '':
            foundIndexTerm = indexWordAlignedAsrTrans
            confidence = whisperTranscriptionDict[file][aligned_asrTrans]['confidence']
            startTime = whisperTranscriptionDict[file][aligned_asrTrans]['start']
            endTime = whisperTranscriptionDict[file][aligned_asrTrans]['end']
        elif indexWordPrompt != '':
            foundIndexTerm = indexWordPrompt
            confidence = whisperTranscriptionDict[file][prompt]['confidence']
            startTime = whisperTranscriptionDict[file][prompt]['start']
            endTime = whisperTranscriptionDict[file][prompt]['end']
        elif indexWordReversedAlignedAsrTrans != '':
            foundIndexTerm = indexWordReversedAlignedAsrTrans
            confidence = whisperTranscriptionDict[file][reversed_aligned_asrTrans]['confidence']
            startTime = whisperTranscriptionDict[file][reversed_aligned_asrTrans]['start']
            endTime = whisperTranscriptionDict[file][reversed_aligned_asrTrans]['end']
        else:
            foundIndexTerm = ''
            confidence = 0
            startTime = 0
            endTime = 0

        foundIndexTermsList.append(foundIndexTerm)
        confidenceList.append(confidence)
        startTimeList.append(startTime)
        endTimeList.append(endTime)

    alignmentDF['whisperWordIndex'] = foundIndexTermsList
    alignmentDF['confidence'] = confidenceList
    alignmentDF['begin'] = startTimeList
    alignmentDF['end'] = endTimeList
    alignmentDF['filename'] = [file] * len(alignmentDF)

    whisperOutputDF = pd.concat([whisperOutputDF, alignmentDF])


def getIndexColumn(row):
    return row['filename'] + '_' + row['prompt']

whisperOutputDF['prompt'] = whisperOutputDF.index
whisperOutputDF['index'] = whisperOutputDF.apply(getIndexColumn, axis=1)
whisperOutputDF = whisperOutputDF.set_index('index')
whisperOutputDF

whisperOutputDF.to_csv('../../astla-data/dart-preposttest/slate-data/5-asr-analysis-results/whisper-word-assessments-all.csv')

## Output/Conclusion
The output of this script can be found here:

    '../../astla-data/dart-preposttest/slate-data/5-asr-analysis-results/whisper-word-assessments-all.csv'

## Some blocks with example code

# EXAMPLE CODE for parsing whisper dictionary

for file in whisperDict.keys():
    data = whisperDict[file]
    print(data.keys())
    # ['text', 'segments', 'language']
    print(data['segments'])
    # list of Segment (Segment[])
    print(data['segments'][0].keys())
    # ['id', 'seek', 'start', 'end', 'text', 'tokens', 'temperature', 'avg_logprob', 'compression_ratio', 'no_speech_prob', 'confidence', 'words']
    print(data['segments'][0]['words'])
    # List of words (Word[])
    print(data['segments'][0]['words'][0]['text'])
    print(data['segments'][0]['words'][0]['start'])
    print(data['segments'][0]['words'][0]['end'])
    print(data['segments'][0]['words'][0]['confidence'])
