==> 6-combine-all-sources.ipynb <==
import pandas as pd

# Binary scores for statistical analysis

def intersection(lst1, lst2):
    lst3 = [value for value in lst1 if value in lst2]
    return lst3


def exception(lst1, lst2):
    lst3 = [x for x in lst2 if x not in lst1]
    return lst3

# Select Novo data and convert to binary scores
novoDF = pd.read_csv(
    '../../astla-data/dart-preposttest/slate-data/5-asr-analysis-results/novo-word-assessments.csv')
binaryNovo = novoDF[['novo_word', 'novo_acc', 'novo_conf']].rename(columns={'novo_word': 'filename'})
binaryNovo['filename'] = binaryNovo['filename'].apply(str.lower)
binaryNovo

# Select allTeacher Data
allTeacherDF = pd.read_csv(
    '../../astla-data/dart-preposttest/slate-data/1-teacherAssessments/all_teacherAssessments.csv')
allTeacherDF['filename'] = allTeacherDF['filename'].apply(str.lower)
allTeacherDF


# Select whisper data and convert to binary scores
whisperDF = pd.read_csv(
    '../../astla-data/dart-preposttest/slate-data/5-asr-analysis-results/whisper-word-assessments-all.csv')
whisperDF['whisper_acc'] = [1 if x == True else 0 for x in whisperDF['correct']]
binaryWhisper = whisperDF[['index', 'whisper_acc', 'confidence']].rename(columns={'index': 'filename'})
binaryWhisper['filename'] = binaryWhisper['filename'].apply(str.lower)
binaryWhisper


phoneticDF = pd.read_csv(
    '../../astla-data/dart-preposttest/slate-data/5-asr-analysis-results/teacherAndManualAssessments.csv')
phoneticDF = phoneticDF[["Filename", "Accuracy", "problem description"]].rename(
    columns={'Filename': 'filename', 'Accuracy': 'mean_teacher_accuracy', 'problem description': 'phonetic_description'})
phoneticDF['filename'] = phoneticDF['filename'].apply(str.lower)
phoneticDF


print(len(set(list(allTeacherDF['filename']))))
print(len(set(list(binaryNovo['filename']))))
print(len(set(list(binaryWhisper['filename']))))
print(len(set(list(phoneticDF['filename']))))


# Merge the three binary dataframes above

# allTeacher + Novo
# binaryTeachersNovoDF = allTeacherDF.merge(
#     right=binaryNovo, left_on='filename', right_on='novo_word')

binaryTeachersNovoDF = pd.merge(allTeacherDF, binaryNovo, on='filename', how='outer')
binaryTeachersNovoWhisperDF = pd.merge(binaryTeachersNovoDF, binaryWhisper, on='filename', how='outer')
binaryTeachersNovoWhisperPhonesDF = pd.merge(
    binaryTeachersNovoWhisperDF, phoneticDF, on='filename', how='outer')
binaryTeachersNovoWhisperPhonesDF.to_csv(
    '/vol/tensusers5/wharmsen/astla-data/dart-preposttest/slate-data/7-binary-assessments/assessments.csv')
