==> 7-create-textgrids-phonetic-trans.ipynb <==
import pandas as pd
import textgrids as tg
import glob

with open('../../astla-data/dart-preposttest/slate-data/6-phonetic-transcriptions/phonetic-lexicon-slate.txt') as f:
    lexiconData = [line.replace('\n', '').split('\t') for line in f.readlines()]

def getPhoneticTranscription(word):
    return pd.DataFrame(lexiconData, columns=['entry', 'phonemes']).set_index('entry').loc[word, 'phonemes'][0]

getPhoneticTranscription('blauw')

pd.DataFrame(lexiconData, columns=['entry', 'phonemes']).set_index('entry').tail(70)

filelist = glob.glob('../../astla-data/dart-preposttest/slate-data/6-phonetic-transcriptions/input-textgrids/*.TextGrid')

for tgfile in filelist:
    file = tg.TextGrid(tgfile)
    print(tgfile)

    # fill prompt tier
    promptPhonemesTier = []
    for interval in file['prompt']:
        currentIntervalText = interval.text
        begin = interval.xmin
        end = interval.xmax

        if(currentIntervalText != ''):
            newIntervalText = getPhoneticTranscription(currentIntervalText)
            print(currentIntervalText, newIntervalText)
        else:
            newIntervalText = ''
            
        promptPhonemesTier.append(tg.Interval(newIntervalText, begin, end))

    file['prompt_phonemes'] = promptPhonemesTier

    file.write(tgfile.replace('input-textgrids', 'output-textgrids'))



def createTextGrid(prompt, teacherAssessment, begin, end):
    

    return file

createTextGrid('lola lala nieur', 'lola(0) lala(1)', 0, 99)
